python-backend/requirements.txt
Flask
Flask-Cors
python-dotenv
pymongo
requests
google-generativeai
uagents
pydantic
polyline